Add files via upload
This was really confusing for me and I got stumped on ladder, and had to cheat a little bit in some of the later problems to figure out what I was doing. If other people feel the same, can we please spend some more time on this? If not, is there some time we can meet so you can walk one of these questions through for me? Thank you :).